Refactoring of dialog module
- Move code out of __init__.py
- Correct spacing of docstrings
- Make use of ' uniform
- Add module level docstring
